Completes exercise 2.1 creating Course component
This commit changes the structure of the react App, making it so that
the App component creates an object that describes the course which
it passes to a new Course component. The object contains the title of
the course and an array of all of the parts of the course. the Course
component renders this information, using the map function to render
each part as a new <p> HTML object.

This commit also separates the Course component into its own js file,
which is imported by App.js.