Completes exercise 2.4 allowing multiple courses
This commit adjusts the App component so that the 'course' object
describing a course is replaced by a 'courses' array, consisting of
multiple objects each of which describes a course. The App component
now returns a <div> consisting of an <h1> with the curriculum name,
followed by a series of Course components generated by using map on
the courses array.

NB: Exercise 2.2 also completed exercise 2.3 (use the reduce method),
and exercise 2.1 also completed exercise 2.5 (put the Course
component in a separate js file and import into App.js).