Completes exercise 2.6 enabling adding names
This commit adds event handler functions to the form to handle text
entry in the name input field and adding names to the persons state
variable by clicking the add button. It also displays the names in
the persons state variable at the bottom of the page through the map
array function.