Completes exercise 2.7 preventing duplicate names
This commit adds functionality to the form submit button handler that
uses the array reduce method to test if the name being submitted is
already a part of the names array. If it is, the handler calls the
alert function to state that the name is already a part of the
phonebook and does not add the name to the names array. Otherwise,
the name is added as in the previous commit.