Displays names in phonebook sorted alphabetically
This commit adds a new function to the App component which maintains
a sorting of the phonebook entries alphabetically by name, assuming
the list begins either empty or sorted. It also changes the check for
repeated entries to leverage the alphaabetical sort so that it runs
in O(log(n)) time, rather than O(n).

Also - this commit addresses an oversight in the solution to exercise
2.7 in the previous commit. The exercise asks for the comparison to
reject duplicate names regardless of capitalization, but that was not
implemented in the previous commit. This commit fixes that, both in
the new comparison function and in the previous comparison (using
reduce, now commented out but still present).