Completes exercises 2.9 and 2.10 adding a filter
This commit adds a filter to the phonebook, which displays only
entries with names that contain the string entered into the filter
textbox. This is case insensitive. A seperate component is added for
the filter, and saved in a separate Filter.js file.

As the phonebook app is separated into multiple components in their
own files, this completes exercise 2.10.