Removes alphabetical sort from phonebook entries
Future exercises for this project have us writing new entries to the
server, rather than just handling them locally. At this point in Full
Stack Open, we have not yet gotten to back-end development, so the
writes would just be handled by default axios json actions. In order
to maintain an alphabetical sort, I would either have to overwrite
the entire server data on each write, or I would have to do a full
sort on the starting data imported from the server - doing either of
these strikes me as too much for a front-end react app; it could be
done, but it just feels like front-end is the wrong place for this
much.

Since the alphabetical sort stuff was mostly a side-task I took on
because it was more efficient big-O-wise, and I'm pretty sure FSO
intends for us to use more straightforward javascript and mapping
functions, I commented out the alphabetical sort functionality and
un-commented the simpler non-sorting code.