Moves add person function to App component
With the alphabetical sort removed, it made more sense to handle the
function for adding a person to the phonebook to the App component,
and pass that function to the PersonForm component. This way the
PersonForm component does not have direct access to the list of
entries, which is outside of its component.