Completes ex. 2.14, adding entry delete buttons
This commit adds a delete button to the Entry component which prompts
the user to confirm, then if confirmed removes the entry from the
phonebook and from the server. The function to enable this is defined
in the App component, then passed to each Entry component to be used
as the onClick for the delete button.