Completes ex. 2.15, enabling phone number updates
This commit changes the behavior when a duplicate name is added so
that it will update the number of the entry, rather than rejecting
the addition. This first prompts the user saying the name is already
in the phonebook and asking to confirm the update, then if confirmed
will make the update on the server and in the phonebook.