Completes ex. 2.16 adding alert messages on action
This commit adds a Notification component which displays alert
messages confirming successful completion of phonebook actions (e.g.,
adding entry, changing number, deleting entry). The messages
have css styling defined in the component and disappear after 2
seconds.

The Notification component takes as a prop a class with 2 members,
one being the text of the alert message and the other being the font
color - this is so for future exercises error messages can be
displayed in a different color.