Preserves most recent alert message for full 3 sec
Previously, each message was accompanied by a set timeout for 3
seconds to reset the message to null. If you performed a new action
during those 3 seconds, the timeout would remove the alert from the
new action after the 3 seconds from the first action.

This commit adds a state variable that is incremented every time a
alert message is set, and decremented every time an alert message 3
second timeout completes. The setTimeout methods now only reset the
alert to null if it is the last alert message (i.e., the state
variable is equal to 1).