Fixes bugs for delete entry message, alert timing
Previously, alert messages for deleting entries were displayed
regardless of whether the user confirmed or canceled the deletion. In
additon, in the previous commit, which aimed to preserve the most
recent alert message for the full 3 seconds even if it overlapped
with a previous message, I neglected to add the relevant code to the
deletion section. This casused me to miss that the new code added in
the previous commit was not using state correctly inside the
setTimeour functions.

This commit first corrects the delete entry alert messages, by moving
the code inside the appropriate if statement. It also corrects the
functionality added in the previous commit, by adding a useRef
variable which points to the messageTimeouts state variable, and is
used inside out the setTimeout functions. This addresses the issue,
in which the setTimeout calls were using the state at the time they
were initialized, not at the end of the timeout.

In addition, this commit does a little bit of code cleanup by adding
a local function handleMessage to the App component which takes as a
parameter a new message object and does all of the necessary steps to
set it as the new message and set its timeout - this code was
repeated in multiple places previously, so making a local function to
handle it makes the code shorter and cleaner.