Completes ex. 2.17, adding error message
This commit catches errors caused when an entry has already been
deleted from the server, but the update is not rendered locally
(e.g., when another window deleted the entry). The commit adds a
catch to the phonebookService.remove call, which displays a red
message saying the entry had already been deleted from the server.