Completes ex. 2.18, creating country data app
This commit creates the basic country data app. It countains a filter
component which creates an input form, and a component that is either
a list of countries which match the filter term or detail for one
specific country when only one country matches. The filtering is
handled by a function defined in the App component and passed as a
prop to the Filter component.