Completes ex. 2.19, adds show detail buttons
This commit creates a Country component for each country listed by
the Countries component when the filter results in 2-10 total
countries. This component includes the country's name and a 'show'
button which, when clicked, displays a CountryInfo component for that
country (when displayed, the button changes to 'hide' which removes
the CountryInfo component when clicked).

Displaying the CountryInfo component and the button's text are both
handled using state variables in the Country component. Specifically,
the Country component includes a state variable after the button that
is set to an empty <div> on initial render and is reset to an empty
<div> when the hide button is clicked. When 'show' is clicked, an
axios call is made to the data server (defined in getOne in
restcountries.js), with the result used to generate the CountryInfo
component.